tests/requirements.txt
selenium==4.7.2
Pillow==9.4.0
pytest==6.2.5
pytest-selenium==4.0.0
